requirements.txt
asgiref==3.2.10
decorator==4.4.2
Django==3.1
django-widget-tweaks==1.4.8
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==1.0.18
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2020.1
simplegeneric==0.8.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
uwsgi
